chore(148/phase1): eliminate warnings by scope-fixing shared test module
Per the new Prime Directive #4 (zero warnings, no underscore hacks), a
warning is a signal — not a chore. When a compiler flags an unused
utility, the correct response is to investigate why it's unused and fix
the root cause, not to silence it with #[allow(dead_code)] or by
renaming the binding to _var.

What went wrong in the previous commits: the Phase 1 tests declared
`mod common;` and pulled in `tests/common/mod.rs`, which contains GCS-
test utilities (TestConfig, get_test_config, print_test_*). My test
binary doesn't use those, so the compiler correctly reported them as
dead in this binary. Rather than fix the mis-scoping, I had added
`#![allow(dead_code)]` to the shared module — silencing warnings for
every future test that pulls in `common`, including cases where a real
unused-code bug might surface.

What this commit does:

* Move the peak-tracking allocator inline into
  tests/test_phase1_zero_copy_assembly.rs. It has exactly one caller
  today; if Phase 2/3/4 needs it later, extract it back then.
* Delete tests/common/peak_alloc.rs.
* Revert the module-level #![allow(dead_code)] on tests/common/mod.rs.
  common/mod.rs returns to its pre-Phase-1 state — only compiled by
  test binaries that actually use its helpers.
* Drop the PeakAlloc::allocations() and reallocations() methods that
  had no callers (YAGNI — add back when a future test genuinely needs
  them, not preemptively).

Also updates CLAUDE.md (project-level) to state the s3dlio-specific
form of the zero-warnings rule, including the specific reasons an
unused binding in this codebase has historically hidden real bugs
(Send/Sync violations, dropped .await chains, unwritten buffer
offsets).

Verified: cargo test --lib -> 325/325, cargo test --test
test_phase1_zero_copy_assembly -> 4/4, cargo fmt --check clean,
cargo clippy --lib --tests --no-deps -- -D warnings clean, and a
fresh compile produces zero warnings from any local crate code.

The parent CLAUDE.md at /home/eval/Documents/Code/CLAUDE.md and the
associated memory files are updated outside this git repo (parent is
not a git tree; memories live in ~/.claude/).
